feat(routines): merge functionality for muscle group filter
The muscle group filter functionality was integrated into the routines page. The Three.js design now displays correctly, though there is still an issue with the filter. Previously, the filter was set up to show routines by muscle group, when it should actually filter by exercise. This issue still needs to be fixed.